[4][fix] - 피드백 받은 부분 fix 1. 컨벤션 적용 2. db 속성 관련 수정에 따른 엔티티 수정 및 구조 수정 3. mailSender return 타입 수정 4. location 유효성 검사 5. 이미지 파일 업로드 api 구현

+++ src/main/java/com/switch_proj/api/api/user/entity/UserLocationEntity.java
@@ -8,15 +8,15 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class UserLocationEntity {
     private Long userLocationId;
-    private Long latitude;
-    private Long longitude;
+    private Float latitude;
+    private Float longitude;
     private String address;
     private Long userId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
     @Builder
-    public UserLocationEntity(Long userLocationId, Long latitude, Long longitude, String address, Long userId, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public UserLocationEntity(Long userLocationId, Float latitude, Float longitude, String address, Long userId, LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.userLocationId = userLocationId;
         this.latitude = latitude;
         this.longitude = longitude;
